api/requirements.txt
fastapi
uvicorn
slowapi
python-dotenv
pyjwt
passlib
bcrypt
python-multipart
email-validator